Initialize pandora sharding info
