+++ validator/client/propose_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/hex"
 	"errors"
-	"github.com/ethereum/go-ethereum/common"
 	"strings"
 	"testing"
 	"time"
@@ -170,7 +169,8 @@ func TestProposeBlock_ProposeBlockFailed(t *testing.T) {
 		gomock.Any(),
 	).Return(testutil.NewBeaconBlock().Block, nil /*err*/)
 
-	header, headerHash, extraData := testutil.NewPandoraBlock(types.Slot(1), 0)
+	header, extraData := testutil.NewPandoraBlock(types.Slot(1), 0)
+	headerHash := sealHash(header)
 	m.pandoraService.EXPECT().GetShardBlockHeader(
 		gomock.Any(), // ctx
 	).Return(header, headerHash, extraData, nil) // nil - error
@@ -236,12 +236,14 @@ func TestProposeBlock_BlocksDoubleProposal(t *testing.T) {
 		gomock.Any(), // epoch
 	).Times(3).Return(&ethpb.DomainResponse{SignatureDomain: make([]byte, 32)}, nil /*err*/)
 
-	header, headerHash, extraData := testutil.NewPandoraBlock(slot, 0)
+	header, extraData := testutil.NewPandoraBlock(slot, 0)
+	headerHash := sealHash(header)
 	m.pandoraService.EXPECT().GetShardBlockHeader(
 		gomock.Any(), // ctx
 	).Return(header, headerHash, extraData, nil) // nil - error
 
-	header, headerHash, extraData = testutil.NewPandoraBlock(slot, 0)
+	header, extraData = testutil.NewPandoraBlock(slot, 0)
+	headerHash = sealHash(header)
 	m.pandoraService.EXPECT().GetShardBlockHeader(
 		gomock.Any(), // ctx
 	).Return(header, headerHash, extraData, nil) // nil - error
@@ -305,12 +307,14 @@ func TestProposeBlock_BlocksDoubleProposal_After54KEpochs(t *testing.T) {
 		gomock.Any(), // epoch
 	).Times(3).Return(&ethpb.DomainResponse{SignatureDomain: make([]byte, 32)}, nil /*err*/)
 
-	header, headerHash, extraData := testutil.NewPandoraBlock(farFuture, 0)
+	header, extraData := testutil.NewPandoraBlock(farFuture, 0)
+	headerHash := sealHash(header)
 	m.pandoraService.EXPECT().GetShardBlockHeader(
 		gomock.Any(), // ctx
 	).Return(header, headerHash, extraData, nil) // nil - error
 
-	header, headerHash, extraData = testutil.NewPandoraBlock(farFuture, 0)
+	header, extraData = testutil.NewPandoraBlock(farFuture, 0)
+	headerHash = sealHash(header)
 	m.pandoraService.EXPECT().GetShardBlockHeader(
 		gomock.Any(), // ctx
 	).Return(header, headerHash, extraData, nil) // nil - error
@@ -358,7 +362,8 @@ func TestProposeBlock_AllowsPastProposals(t *testing.T) {
 		gomock.Any(),
 	).Return(blk.Block, nil /*err*/)
 
-	header, headerHash, extraData := testutil.NewPandoraBlock(farAhead, 0)
+	header, extraData := testutil.NewPandoraBlock(farAhead, 0)
+	headerHash := sealHash(header)
 	m.pandoraService.EXPECT().GetShardBlockHeader(
 		gomock.Any(), // ctx
 	).Return(header, headerHash, extraData, nil) // nil - error
@@ -391,7 +396,8 @@ func TestProposeBlock_AllowsPastProposals(t *testing.T) {
 		gomock.Any(),
 	).Return(blk2.Block, nil /*err*/)
 
-	header, headerHash, extraData = testutil.NewPandoraBlock(past, 0)
+	header, extraData = testutil.NewPandoraBlock(past, 0)
+	headerHash = sealHash(header)
 	m.pandoraService.EXPECT().GetShardBlockHeader(
 		gomock.Any(), // ctx
 	).Return(header, headerHash, extraData, nil) // nil - error
@@ -431,7 +437,8 @@ func TestProposeBlock_AllowsSameEpoch(t *testing.T) {
 		gomock.Any(),
 	).Return(blk.Block, nil /*err*/)
 
-	header, headerHash, extraData := testutil.NewPandoraBlock(farAhead, 0)
+	header, extraData := testutil.NewPandoraBlock(farAhead, 0)
+	headerHash := sealHash(header)
 	m.pandoraService.EXPECT().GetShardBlockHeader(
 		gomock.Any(), // ctx
 	).Return(header, headerHash, extraData, nil) // nil - error
@@ -463,7 +470,8 @@ func TestProposeBlock_AllowsSameEpoch(t *testing.T) {
 		gomock.Any(),
 	).Return(blk2.Block, nil /*err*/)
 
-	header, headerHash, extraData = testutil.NewPandoraBlock(blk2.Block.Slot, 0)
+	header, extraData = testutil.NewPandoraBlock(blk2.Block.Slot, 0)
+	headerHash = sealHash(header)
 	m.pandoraService.EXPECT().GetShardBlockHeader(
 		gomock.Any(), // ctx
 	).Return(header, headerHash, extraData, nil) // nil - error
@@ -495,7 +503,8 @@ func TestProposeBlock_BroadcastsBlock(t *testing.T) {
 		gomock.Any(),
 	).Return(testutil.NewBeaconBlock().Block, nil /*err*/)
 
-	header, headerHash, extraData := testutil.NewPandoraBlock(types.Slot(1), 0)
+	header, extraData := testutil.NewPandoraBlock(types.Slot(1), 0)
+	headerHash := sealHash(header)
 	m.pandoraService.EXPECT().GetShardBlockHeader(
 		gomock.Any(), // ctx
 	).Return(header, headerHash, extraData, nil) // nil - error
@@ -540,7 +549,8 @@ func TestProposeBlock_BroadcastsBlock_WithGraffiti(t *testing.T) {
 		gomock.Any(),
 	).Return(blk.Block, nil /*err*/)
 
-	header, headerHash, extraData := testutil.NewPandoraBlock(types.Slot(1), 0)
+	header, extraData := testutil.NewPandoraBlock(types.Slot(1), 0)
+	headerHash := sealHash(header)
 	m.pandoraService.EXPECT().GetShardBlockHeader(
 		gomock.Any(), // ctx
 	).Return(header, headerHash, extraData, nil) // nil - error
@@ -893,146 +903,4 @@ func TestGetGraffitiOrdered_Ok(t *testing.T) {
 		require.NoError(t, err)
 		require.DeepEqual(t, want, got)
 	}
-}
-
-// TestVerifyPandoraHeader_Ok method checks pandora header validation method
-func TestVerifyPandoraShardHeader(t *testing.T) {
-	validator, _, _, finish := setup(t)
-	defer finish()
-
-	blk := testutil.NewBeaconBlock()
-	blk.Block.Slot = 98
-	blk.Block.ProposerIndex = 23
-	epoch := types.Epoch(uint64(blk.Block.Slot) / 32)
-
-	header, headerHash, extraData := testutil.NewPandoraBlock(blk.Block.Slot, uint64(blk.Block.ProposerIndex))
-
-	// Checks all the validations
-	err := validator.verifyPandoraShardHeader(blk.Block, blk.Block.Slot, epoch, header, headerHash, extraData)
-	require.NoError(t, err, "Should pass without any error")
-
-	// Should get an `errInvalidHeaderHash` error
-	header.Time = uint64(14265167)
-	want := "invalid header hash"
-	err = validator.verifyPandoraShardHeader(blk.Block, blk.Block.Slot, epoch, header, headerHash, extraData)
-	require.ErrorContains(t, want, err, "Should get an errInvalidHeaderHash error")
-
-	// Should get an `errInvalidSlot` error
-	header.Time = uint64(1426516743)
-	blk.Block.Slot = 90
-	want = "invalid slot"
-	err = validator.verifyPandoraShardHeader(blk.Block, blk.Block.Slot, epoch, header, headerHash, extraData)
-	require.ErrorContains(t, want, err, "Should get an errInvalidSlot error")
-
-	// Should get an `errInvalidEpoch` error
-	blk.Block.Slot = 98
-	epoch = 2
-	want = "invalid epoch"
-	err = validator.verifyPandoraShardHeader(blk.Block, blk.Block.Slot, epoch, header, headerHash, extraData)
-	require.ErrorContains(t, want, err, "Should get an errInvalidEpoch error")
-
-	// Shoud get an `errInvalidProposerIndex` error
-	epoch = 3
-	blk.Block.ProposerIndex = 190
-	want = "invalid proposer index"
-	err = validator.verifyPandoraShardHeader(blk.Block, blk.Block.Slot, epoch, header, headerHash, extraData)
-	require.ErrorContains(t, want, err, "Should get an errInvalidProposerIndex error")
-}
-
-// TestProcessPandoraShardHeader method checks the `processPandoraShardHeader`
-func TestProcessPandoraShardHeader(t *testing.T) {
-	validator, m, _, finish := setup(t)
-	defer finish()
-
-	secretKey, err := bls.SecretKeyFromBytes(bytesutil.PadTo([]byte{1}, 32))
-	require.NoError(t, err, "Failed to generate key from bytes")
-	publicKey := secretKey.PublicKey()
-	var pubKey [48]byte
-	copy(pubKey[:], publicKey.Marshal())
-	km := &mockKeymanager{
-		keysMap: map[[48]byte]bls.SecretKey{
-			pubKey: secretKey,
-		},
-	}
-	validator.keyManager = km
-	blk := testutil.NewBeaconBlock()
-	blk.Block.Slot = 98
-	blk.Block.ProposerIndex = 23
-	epoch := types.Epoch(uint64(blk.Block.Slot) / 32)
-
-	// Check with happy path
-	header, headerHash, extraData := testutil.NewPandoraBlock(blk.Block.Slot, uint64(blk.Block.ProposerIndex))
-	m.pandoraService.EXPECT().GetShardBlockHeader(
-		gomock.Any(), // ctx
-	).Return(header, headerHash, extraData, nil) // nil - error
-
-	m.pandoraService.EXPECT().SubmitShardBlockHeader(
-		gomock.Any(), // ctx
-		gomock.Any(), // blockNonce
-		gomock.Any(), // headerHash
-		gomock.Any(), // sig
-	).Return(true, nil)
-
-	status, err := validator.processPandoraShardHeader(context.Background(), blk.Block, blk.Block.Slot, epoch, pubKey)
-	require.NoError(t, err, "Should sucess")
-	require.Equal(t, true, status, "Pandora shrad processing should be successful")
-
-	// Return rlp decoding error when calls `GetWork` api
-	ErrRlpDecoding := errors.New("rlp: input contains more than one value")
-	m.pandoraService.EXPECT().GetShardBlockHeader(
-		gomock.Any(), // ctx
-	).Return(nil, common.Hash{}, nil, ErrRlpDecoding)
-	_, err = validator.processPandoraShardHeader(context.Background(), blk.Block, blk.Block.Slot, epoch, pubKey)
-	require.ErrorContains(t, "rlp: input contains more than one value", ErrRlpDecoding)
-}
-
-// TestValidator_ProposeBlock_Failed_WhenSubmitShardInfoFails methods checks when `SubmitShardInfo` fails
-func TestValidator_ProposeBlock_Failed_WhenSubmitShardInfoFails(t *testing.T) {
-	hook := logTest.NewGlobal()
-	validator, m, validatorKey, finish := setup(t)
-	defer finish()
-	pubKey := [48]byte{}
-	copy(pubKey[:], validatorKey.PublicKey().Marshal())
-
-	m.validatorClient.EXPECT().DomainData(
-		gomock.Any(), // ctx
-		gomock.Any(), // epoch
-	).Times(2).Return(&ethpb.DomainResponse{SignatureDomain: make([]byte, 32)}, nil /*err*/)
-
-	m.validatorClient.EXPECT().GetBlock(
-		gomock.Any(), // ctx
-		gomock.Any(),
-	).Times(2).Return(testutil.NewBeaconBlock().Block, nil /*err*/)
-
-	header, headerHash, extraData := testutil.NewPandoraBlock(types.Slot(1), 0)
-	m.pandoraService.EXPECT().GetShardBlockHeader(
-		gomock.Any(), // ctx
-	).Return(header, headerHash, extraData, nil) // nil - error
-
-	header, headerHash, extraData = testutil.NewPandoraBlock(types.Slot(1), 0)
-	m.pandoraService.EXPECT().GetShardBlockHeader(
-		gomock.Any(), // ctx
-	).Return(header, headerHash, extraData, nil) // nil - error
-
-	// When `SubmitShardInfo` api returns false status
-	m.pandoraService.EXPECT().SubmitShardBlockHeader(
-		gomock.Any(), // ctx
-		gomock.Any(), // blockNonce
-		gomock.Any(), // headerHash
-		gomock.Any(), // sig
-	).Return(false, nil)
-
-	validator.ProposeBlock(context.Background(), 1, pubKey)
-	require.LogsContain(t, hook, "Failed to process pandora chain shard header")
-
-	// When `SubmitShardInfo` api returns error
-	m.pandoraService.EXPECT().SubmitShardBlockHeader(
-		gomock.Any(), // ctx
-		gomock.Any(), // blockNonce
-		gomock.Any(), // headerHash
-		gomock.Any(), // sig
-	).Return(true, errors.New("Failed to submit shard header"))
-
-	validator.ProposeBlock(context.Background(), 1, pubKey)
-	require.LogsContain(t, hook, "Failed to process pandora chain shard header")
 }
